scripts/requirements.txt
telethon
aiohttp
supabase
python-dotenv
cryptg
